Revamped msetavatarlibrary to mpushavatartolibrary. Will remove oldest avatar if avatar limit is exceeded
