+++ src/sockets/commands/mod/mpushavatartolibrary.ts
@@ -1,0 +1,60 @@
+import { Command } from '../types';
+import { sendReplyToCommand } from '../../sockets';
+import { SocketUser } from '../../types';
+import User from '../../../models/user';
+import { S3Agent } from '../../../s3/S3Agent';
+import S3Controller from '../../../s3/S3Controller';
+
+export const mpushavatartolibrary: Command = {
+  command: 'mpushavatartolibrary',
+  help: "/mpushavatartolibrary <player name> <avatar id>: Add an approved avatar ID to a user's avatar library. It will remove the oldest avatar if the max limit has been exceeded.",
+  async run(args: string[], senderSocket: SocketUser) {
+    if (args.length !== 3) {
+      sendReplyToCommand(
+        senderSocket,
+        'Please specify <username> <avatar id>. Example: /mpushavatartolibrary asdf 1',
+      );
+      return;
+    }
+
+    const usernameLower = args[1].toLowerCase();
+    const user = await User.findOne({ usernameLower });
+
+    if (!user) {
+      sendReplyToCommand(
+        senderSocket,
+        `Invalid username. Could not find: ${usernameLower}`,
+      );
+      return;
+    }
+
+    const toBeAddedAvatarId = Number(args[2]);
+    const s3Agent = new S3Agent(new S3Controller());
+    const approvedIds = await s3Agent.getApprovedAvatarIdsForUser(
+      usernameLower,
+    );
+
+    if (isNaN(toBeAddedAvatarId) || !approvedIds.includes(toBeAddedAvatarId)) {
+      sendReplyToCommand(
+        senderSocket,
+        `Invalid avatar ID received. List of valid IDs: ${approvedIds}`,
+      );
+      return;
+    }
+
+    user.avatarLibrary.push(toBeAddedAvatarId);
+    // TODO-kev: Add a check for max library size here. If greater than max, remove oldest approved avatar
+    // Ideally will not be a cyclic nature. Consider mreplaceavatar
+    // if (user.avatarLibrary.length > MAX) {
+    //   user.avatarLibrary.shift();
+    // }
+    user.markModified('avatarLibrary');
+    await user.save();
+
+    sendReplyToCommand(
+      senderSocket,
+      `Successfully updated ${usernameLower}'s avatar library: ${user.avatarLibrary}`,
+    );
+    return;
+  },
+};
